DATA605/Spring2025/projects/TutorTask300_Spring2025_Real_Time_Bitcoin_Data_Processing_with_BigDL/requirements.txt
bigdl-spark
pyspark
requests
matplotlib
